Refactor utility imports to use consistent naming convention and add new utility functions for organization and vendor risk management
